Started moving hooks from app file
